feat: add global exception handling and validation errors

+++ src/main/java/com/devsuperior/dsclient/services/ClientService.java
@@ -3,8 +3,11 @@ package com.devsuperior.dsclient.services;
 import com.devsuperior.dsclient.dto.ClientDTO;
 import com.devsuperior.dsclient.entities.Client;
 import com.devsuperior.dsclient.repositories.ClientRepository;
+import com.devsuperior.dsclient.services.exceptions.DatabaseException;
+import com.devsuperior.dsclient.services.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -18,7 +21,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
-        Client client = clientRepository.findById(id).get();
+        Client client = clientRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ClientDTO(client);
     }
 
@@ -38,15 +42,23 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
+        try {
             Client entity = clientRepository.getReferenceById(id);
             copyDtoToEnity(dto, entity);
             entity = clientRepository.save(entity);
             return new ClientDTO(entity);
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }
     }
 
     @Transactional
     public void delete(Long id) {
-        clientRepository.deleteById(id);
+        if (!clientRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }
+            clientRepository.deleteById(id);
     }
 
     private void copyDtoToEnity(ClientDTO dto, Client entity) {
